+++ kinto/core/storage/testing.py
@@ -476,6 +476,28 @@ class BaseTestStorage:
                                           **self.storage_kw)
         self.assertEqual(len(records), 1)
 
+    def test_get_all_can_filter_on_array_with_contains_and_unsupported_type(self):
+        self.create_record({'code': 'black'})
+        self.create_record({'fib': [2, 3, 5]})
+        self.create_record({'fib': [3, 5, 8]})
+        self.create_record({'fib': [5, 8, 13]})
+
+        filters = [Filter('fib', [{"demo": "foobar"}], utils.COMPARISON.CONTAINS)]
+        records, _ = self.storage.get_all(filters=filters,
+                                          **self.storage_kw)
+        self.assertEqual(len(records), 0)
+
+    def test_get_all_can_filter_on_array_with_contains_any_and_unsupported_type(self):
+        self.create_record({'code': 'black'})
+        self.create_record({'fib': [2, 3, 5]})
+        self.create_record({'fib': [3, 5, 8]})
+        self.create_record({'fib': [5, 8, 13]})
+
+        filters = [Filter('fib', [{"demo": "foobar"}], utils.COMPARISON.CONTAINS_ANY)]
+        records, _ = self.storage.get_all(filters=filters,
+                                          **self.storage_kw)
+        self.assertEqual(len(records), 0)
+
     def test_get_all_can_filter_with_numeric_values(self):
         self.create_record({'missing': 'code'})
         for l in [1, 10, 6, 46]: